+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -125,7 +125,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01JVPYM2BBQB2BVWSVQ0WEEMJ9"
+        "01JVRRQ62Q3CWC8YKGXT9GSGNN"
     )
     r"""The id of the resource"""