+++ packages/orq-rc/RELEASES.md
@@ -1339,3 +1339,13 @@ Based on:
 - [python v3.6.0-rc.19] packages/orq-rc
 ### Releases
 - [PyPI v3.6.0-rc.19] https://pypi.org/project/orq-ai-sdk/3.6.0-rc.19 - packages/orq-rc
+
+## 2025-05-21 06:34:27
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.548.3 (2.605.6) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v3.6.0-rc.20] packages/orq-rc
+### Releases
+- [PyPI v3.6.0-rc.20] https://pypi.org/project/orq-ai-sdk/3.6.0-rc.20 - packages/orq-rc